reducing api surface by moving date utils and some of the card utils functionality
